review comments incoporated.
