--- modules/integration/test-results-service/dependency-reduced-pom.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-  <parent>
-    <artifactId>msf4j-service</artifactId>
-    <groupId>org.wso2.msf4j</groupId>
-    <version>2.1.0</version>
-  </parent>
-  <modelVersion>4.0.0</modelVersion>
-  <groupId>org.wso2.sp.events</groupId>
-  <artifactId>test-results-service</artifactId>
-  <name>WSO2 MSF4J Microservice</name>
-  <version>1.0.0-SNAPSHOT</version>
-  <properties>
-    <microservice.mainClass>org.wso2.sp.events.Application</microservice.mainClass>
-  </properties>
-</project>
-
