+++ modules/integration/test-results-service/pom.xml
@@ -20,17 +20,15 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 
     <parent>
-        <groupId>org.wso2.msf4j</groupId>
-        <artifactId>msf4j-service</artifactId>
-        <version>2.1.0</version>
+        <groupId>org.wso2.sp</groupId>
+        <artifactId>wso2sp-integration</artifactId>
+        <version>4.0.0-SNAPSHOT</version>
+        <relativePath>../pom.xml</relativePath>
     </parent>
     <modelVersion>4.0.0</modelVersion>
-
-    <groupId>org.wso2.sp.events</groupId>
     <artifactId>test-results-service</artifactId>
     <version>1.0.0-SNAPSHOT</version>
-    <name>WSO2 MSF4J Microservice</name>
-
+    <name>Integration Test- Siddhi Events Service</name>
 
     <dependencies>
     <!-- Uncomment the following if you want to enable metrics & monitoring -->
@@ -55,10 +53,42 @@
             <artifactId>jackson-jaxrs</artifactId>
             <version>1.1.1</version>
         </dependency>
+        <dependency>
+            <groupId>org.wso2.msf4j</groupId>
+            <artifactId>msf4j-core</artifactId>
+        </dependency>
     </dependencies>
-
+    <build>
+        <plugins>
+        <plugin>
+            <groupId>org.apache.maven.plugins</groupId>
+            <artifactId>maven-assembly-plugin</artifactId>
+            <executions>
+                <execution>
+                    <phase>package</phase>
+                    <goals>
+                        <goal>single</goal>
+                    </goals>
+                </execution>
+            </executions>
+            <configuration>
+                <archive>
+                    <manifest>
+                        <addClasspath>true</addClasspath>
+                        <mainClass>org.wso2.sp.tests.eventscollector.Application</mainClass>
+                    </manifest>
+                </archive>
+                <descriptorRefs>
+                    <descriptorRef>jar-with-dependencies</descriptorRef>
+                </descriptorRefs>
+                <appendAssemblyId>false</appendAssemblyId>
+            </configuration>
+        </plugin>
+        </plugins>
+    </build>
     <properties>
-        <microservice.mainClass>org.wso2.sp.events.Application</microservice.mainClass>
+    <microservice.mainClass>org.wso2.sp.tests.eventscollector.Application</microservice.mainClass>
+	<maven.findbugsplugin.version.exclude>findbugs-exclude.xml</maven.findbugsplugin.version.exclude>
     </properties>
 
 </project>
